NVIDIA: SAUCE: HACK: fs/resctrl: Add cranky debug for reading CPU msr values
BugLink: https://bugs.launchpad.net/bugs/2122432

MPAM affects performance, which makes it hard to debug. (is that the
performance I was expecting?). One source of confusion is whether a
task is having its traffic labelled with the correct PARTID. This
is fairly straightforward for tasks, but once the resctrl 'cpus' files
get involved, the picture is more confusing.

As a debug aid, add a file in /proc that uses task_work to read the
CPU register for a task. This is output raw without interpretation.

This is not for upstream.

task_work may not be able to complete its work quickly if the task
is unable to take signals. Don't be too surprised if for some tasks
this appears to block the readers of the file.

Signed-off-by: James Morse <james.morse@arm.com>
(cherry picked from commit 8808b4be0ff62853e802eda0615bd6b73a8b2ec8 https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git)
Signed-off-by: Fenghua Yu <fenghuay@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Signed-off-by: Ian May <ianm@nvidia.com>