net/mlx5: Generate PPS IN event on new function for shared clock
BugLink: https://bugs.launchpad.net/bugs/2124107

As a specific function (mdev) is chosen to send MTPPSE command to
firmware, the event is generated only on that function. When that
function is unloaded, the PPS event can't be forward to PTP device,
even when there are other functions in the group, and PTP device is
not destroyed. To resolve this problem, need to send MTPPSE again from
new function, and dis-arm the event on old function after that.

PPS events are handled by EQ notifier. The async EQs and notifiers are
destroyed in mlx5_eq_table_destroy() which is called before
mlx5_cleanup_clock(). During the period between
mlx5_eq_table_destroy() and mlx5_cleanup_clock(), the events can't be
handled. To avoid event loss, add mlx5_clock_unload() in mlx5_unload()
to arm the event on other available function, and mlx5_clock_load in
mlx5_load() for symmetry.

Signed-off-by: Jianbo Liu <jianbol@nvidia.com>
Reviewed-by: Carolina Jubran <cjubran@nvidia.com>
Reviewed-by: Dragos Tatulea <dtatulea@nvidia.com>
Signed-off-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
(cherry picked from commit 39c1202fa9428bcb8d1242ee12f81cbcb298c020)
Signed-off-by: Benjamin Poirier <bpoirier@nvidia.com>
Acked-by: Valentine Fatiev <valentinef@nvidia.com>
Acked-by: Aya Levin <ayal@nvidia.com>
Tested-by: Rola Dabbah <rdabbah@nvidia.com>
Tested-by: Shmuel Gutman <sgutman@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Kai-Heng Feng <kaihengf@nvidia.com>
Signed-off-by: Ian May <ianm@nvidia.com>