net/mlx5: HWS, fix redundant extension of action templates
BugLink: https://bugs.launchpad.net/bugs/2124107

When a rule is inserted into a matcher, we search for the suitable
action template. If such template is not found, action template array
is extended with the new template. However, when several threads are
performing this in parallel, there is a race - we can end up with
extending the action templates array with the same template.

This patch is doing the following:
 - refactor the code to find action template index in rule create and
   update, have the common code in an auxiliary function
 - after locking all the queues, check again if the action template
   array still needs to be extended

Signed-off-by: Vlad Dogaru <vdogaru@nvidia.com>
Signed-off-by: Yevgeny Kliteynik <kliteyn@nvidia.com>
Reviewed-by: Mark Bloch <mbloch@nvidia.com>
Signed-off-by: Tariq Toukan <tariqt@nvidia.com>
Link: https://patch.msgid.link/1746992290-568936-9-git-send-email-tariqt@nvidia.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 041861b40f599311214a52075140db8be29fd27f)
Signed-off-by: Benjamin Poirier <bpoirier@nvidia.com>
Acked-by: Valentine Fatiev <valentinef@nvidia.com>
Acked-by: Amir Tzin <amirtz@nvidia.com>
Tested-by: Rola Dabbah <rdabbah@nvidia.com>
Tested-by: Shmuel Gutman <sgutman@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Kai-Heng Feng <kaihengf@nvidia.com>
Signed-off-by: Ian May <ianm@nvidia.com>