net/mlx5: Remove default QoS group and attach vports directly to root TSAR
BugLink: https://bugs.launchpad.net/bugs/2124107

Currently, the driver creates a default group (`node0`) and attaches
all vports to it unless the user explicitly sets a parent group. As a
result, when a user configures tx_share on a group and tx_share on
a VF, the expectation is for the group and the VF to share bandwidth
relatively. However, since the VF is not connected to the same parent
(but to the default node), the proportional share logic is not applied
correctly.

To fix this, remove the default group (`node0`) and instead connect
vports directly to the root TSAR when no parent is specified. This
ensures that vports and groups share the same root scheduler and their
tx_share values are compared directly under the same hierarchy.

Fixes: 0fe132eac38c ("net/mlx5: E-switch, Allow to add vports to rate groups")
Signed-off-by: Carolina Jubran <cjubran@nvidia.com>
Reviewed-by: Cosmin Ratiu <cratiu@nvidia.com>
Signed-off-by: Mark Bloch <mbloch@nvidia.com>
Link: https://patch.msgid.link/20250820133209.389065-3-mbloch@nvidia.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 330f0f6713a39581936decac72331e6ab7f13529)
Signed-off-by: Benjamin Poirier <bpoirier@nvidia.com>
Acked-by: Valentine Fatiev <valentinef@nvidia.com>
Acked-by: Amir Tzin <amirtz@nvidia.com>
Tested-by: Rola Dabbah <rdabbah@nvidia.com>
Tested-by: Shmuel Gutman <sgutman@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Kai-Heng Feng <kaihengf@nvidia.com>
Signed-off-by: Ian May <ianm@nvidia.com>